feat: Slice 4 - Available Inventory & Net Requirement
- Domain: IngredientRequirement value object with demand/available/net quantities
- Domain: calculate_inventory_requirements() - compares demand vs inventory with unit conversion
- Application: calculate_weekly_inventory_requirements() + get_inventory_requirements_for_week()
- Presentation: /requirements and /requirements/<week_start> routes with templates
- Tests: 26 new tests (domain + architecture) - 175 total tests pass
- Architecture: No NetRequirement/AvailableInventory tables, no persistence, no inventory mutation
- No new migration needed (derived calculation only)

Fixes:
- Fixed UnitSystem to use Decimal for precise arithmetic (eliminated float conversions)
- Fixed Ingredient domain to use Decimal for quantity
- Fixed IngredientDemand immutability (removed object.__setattr__ bypass)
- Fixed calculate_demand_for_week to return correct result (was returning demands instead of result)
- Updated architecture tests for new Slice 3/4 domain modules
- Added comprehensive Slice 4 tests (domain, persistence, application, architecture)

+++ dinner_spinner/domain/ingredient.py
@@ -11,6 +11,7 @@ Ingredient represents the household's current inventory state.  The Ingredient
 itself is the inventory holding.  There is no separate Container entity in V1.
 """
 
+from decimal import Decimal
 from dinner_spinner.domain.unit_system import validate_unit, is_initialized, convert, category_of
 
 
@@ -21,7 +22,7 @@ class Ingredient:
     """
 
     def __init__(self, id: int, name: str, inventory_category_id: int | None,
-                 quantity: float, unit: str):
+                 quantity: Decimal, unit: str):
         if not name or not name.strip():
             raise ValueError("Ingredient name is required")
         self.id = id
@@ -63,7 +64,7 @@ class Ingredient:
         return hash((self.id, self.name, self.inventory_category_id,
                      self.quantity, self.unit))
 
-    def increase_quantity(self, quantity: float, unit: str) -> None:
+    def increase_quantity(self, quantity: Decimal, unit: str) -> None:
         """Increase the ingredient's quantity by the given amount in the specified unit.
 
         The quantity is converted to the ingredient's current unit using the UnitSystem
@@ -102,7 +103,7 @@ class Ingredient:
         converted_qty = convert(quantity, unit, self.unit)
         self.quantity += converted_qty
 
-    def decrease_quantity(self, quantity: float, unit: str) -> None:
+    def decrease_quantity(self, quantity: Decimal, unit: str) -> None:
         """Decrease the ingredient's quantity by the given amount in the specified unit.
 
         The quantity is converted to the ingredient's current unit using the UnitSystem
@@ -114,7 +115,7 @@ class Ingredient:
 
         Raises:
             ValueError: If quantity <= 0, unit invalid, unit incompatible,
-                       or resulting quantity would be negative
+                        or resulting quantity would be negative
             RuntimeError: If UnitSystem not initialized
         """
         if quantity <= 0:
